TizenViewElementary: Change the size of view when the size of container changes
When the size of the container is changed from the outside,
the size of FlutterView and the size of the image are changed together.

+)
Basically, the size of the container is set according to the size of the image.
It was enough to set min/max size only for images.
After setting the image size for an unknown reason,
the container resize callback is called once again.
At this time, evas_object_geometry_get() of container_ returns
the initially set value and the size backs to the original size.
Therefore, we set min/max of container_ together.