Propagation of unhandled key event of TizenView
When flutter's OnKeyEvent(OnKey) returns KeyEventResult.ignored,
it passes the KeyEvent to the upper level of FlutterView.

In case of arrowkey(up, down, left, right), handle always returns true
when KeyDown event regardless of the return value.
Therefore, in this case, the handle value of the KeyUp event is checked
and a feed event is generated.